+++ sharedPlugins/clusterioMod/config.js
@@ -5,6 +5,5 @@ module.exports = {
 	// Name of package. For display somewhere I guess.
 	name: "clusterioMod",
 	version: "1.0.0",
-	binary: "nodePackage",
 	description: "Server side handling of functionality in the clusterio lua mod",
 }