+++ sharedPlugins/remoteCommands/config.js
@@ -8,6 +8,5 @@ module.exports = {
 	// Name of package. For display somewhere I guess.
 	name: "remoteCommands",
 	version: "1.0.0",
-	binary: "nodePackage",
 	description: "Config for clusterio remote command plugin. Allows you to execute commands on slaves with a web API.",
 }