+++ sharedPlugins/serverManager/config.js
@@ -4,7 +4,6 @@
 module.exports = {
 	name: "serverManager",
 	version: "1.0.0",
-	binary: "nodePackage",
 	description: "Server management tools for remote config edits and web based server management",
 	masterPlugin: "masterPlugin.js",
 	dependencies: [